Fix Demo Site Pages The legacyPagesRouterSitePreviewApiHandler used parts from SitePreviewUtils which has a server-only restriction. The server-only package can only be used in RSC compatible setups (i.e. the App Router). To fix this, we move the App Router specific parts to a separate file, ensuring that SitePreviewUtils can be used for both Pages Router and App Router.
